updated tables list

+++ Client/TawFrontend/src/app/tables-list/tables-list.component.ts
@@ -11,6 +11,7 @@ import {is} from "date-fns/locale";
 })
 export class TablesListComponent implements OnInit {
   tables: Table[] = [];
+  showOptions: boolean = false;
 
   constructor(private tableService: TableService, private socketIo: SocketioService) {
   }
@@ -33,17 +34,31 @@ export class TablesListComponent implements OnInit {
   }
 
   deleteTable(tableNum: number, isFree: boolean) {
-    if(!isFree){
-      alert("You can't delete a table occupied!");
+    if(confirm('Do you really want to delete table '+tableNum+' ?')) {
+      if (!isFree) {
+        alert("You can't delete a table occupied!");
+        return;
+      }
+      this.tableService.deleteTable(tableNum).subscribe({
+        next: (res) => console.log('Table ' + tableNum + 'deleted'),
+        error: (err) => console.log('Error deleting table')
+      });
+    }
+  }
+  createTable(tableNum: string, seats: string){
+    if(tableNum == '' || seats == ''){
+      alert('Params are not correct');
       return;
     }
-    this.tableService.deleteTable(tableNum).subscribe({
-      next:(res)=> console.log('Table '+ tableNum + 'deleted'),
-      error:(err) => console.log('Error deleting table')
+    let tn = parseInt(tableNum, 10);
+    let s = parseInt(seats, 10);
+    this.tableService.createTable(tn, s).subscribe({
+      next:(res)=> {
+        console.log('Table ' + tableNum + 'created');
+        this.showOptions = false;
+      },
+      error:(err) => console.log('Error creating table')
     });
-  }
-  createTable(tableNum: number, seats: number){
-
   }
 
 }
